+++ NEWS.md
@@ -1,9 +1,12 @@
 # brickster 0.2.7
+
 * Exporting UC table functions (`db_uc_table*`) (#72)
 * Adding support for `direct_download` option in `db_workspace_export()`
 * Exporting UC Catalog/Schema get/list functions (#72)
+* Adding support for UC Volume management (#72)
 
 # brickster 0.2.6
+
 * Fixing `db_volume_delete()` function (#73, @vladimirobucina)
 * Adjustments to ensure {httr2} changes don't break things (#75, @hadley)